Datamanager's read() now supports unique reading. Also changed the delete function to also use unique tags only (basically you won't accidentally delete stuff)

+++ datamanager.py
@@ -61,17 +61,23 @@ class Datamanager():
             return []
         #Find all rows in association matching all given tags
         if unique:
-            print 'Unique not properly implemented yet'
-            '''relation_subq = self.s.query(association_table).\
+            #Because of how queries work I split this query into 2 to make sure that
+            #entries with more tags than the given amount would not match
+            ids = self.s.query(association_table.c.meas_id).\
                     join(Tag).filter(Tag.tag.in_(tags)).\
-                    subquery()
-            print 'matching unique'
-
-            matches = self.s.query(Measurement).join(relation_subq).\
                     group_by(association_table.c.meas_id).\
+                    having(func.count(distinct(association_table.c.tag_id)) >= len(tags)).\
+                    all()
+            new_ids = []
+            for id in ids:
+                new_ids.append(id[0])
+
+            #This query will filter out any row with more tags than the number of given ones.
+            matches = self.s.query(Measurement).filter(Measurement.id.in_(new_ids)).\
+                    join(association_table).\
+                    group_by(Measurement.id).\
                     having(func.count(distinct(association_table.c.tag_id)) == len(tags)).\
                     all()
-            '''
         else:
             relation_subq = self.s.query(association_table.c.meas_id).\
                     join(Tag).filter(Tag.tag.in_(tags)).\
@@ -87,7 +93,7 @@ class Datamanager():
         return data_objects
 
     """
-    Delete all matching measurements, tags are not removed but may not be associated with
+    Delete all uniquely matching measurements, tags are not removed but may not be associated with
     any measurements
     """
     def delete(self, tags):
@@ -106,6 +112,18 @@ class Datamanager():
         for i, id in enumerate(ids):
             ids[i] = ids[i][0]
 
+        #Sort out the non/unique matches
+        ids = self.s.query(Measurement.id).filter(Measurement.id.in_(ids)).\
+                join(association_table).\
+                group_by(Measurement.id).\
+                having(func.count(distinct(association_table.c.tag_id)) == len(tags)).\
+                all()
+        for i, id in enumerate(ids):
+            ids[i] = ids[i][0]
+
+        if len(ids) == 0:
+            print 'No matches to delete'
+            return
         self.s.execute(Measurement.__table__.delete().where(Measurement.id.in_(ids)))
         self.s.execute(association_table.delete().where(association_table.c.meas_id.in_(ids)))
         self.s.commit()
@@ -140,7 +158,7 @@ class Datamanager():
                 if new:
                     combinations.append([tags, 1])
 
-                #Startes the next entry
+                #Starts the next entry
                 index = row[0]
                 tags = [row[1]]
 
@@ -180,7 +198,7 @@ class Data():
 if __name__ == '__main__':
     dm = Datamanager(echo=True)
     dm.list_combinations()
-    #dm.delete(['valid100'])
+    #print dm.read(['sgt50'], unique=True)
     #dm.insert(tags=['sgt', 'training', 'test'], observable=100, energy=50)
     #dm.insert(tags=['sgt', 'training'], observable=20, energy=40)
     #dm.insert(tags=['sgt', 'training'], observable=10, energy=5)
